Update stock_orderpoint.py
related to this ticket: https://www.odoo.com/web#id=3790957&cids=17&menu_id=4720&action=333&active_id=49&model=project.task&view_type=form

Automatic Reordering rules are created when the forecast of a product is negative (right behaviour) 

For customers can be confusing looking at a product and finding they have a reordering rule, setting new min-max rules and then it disappears. 

The pull I´m proposing is to add the write_uid (last modified by) parameter to make sure that if a customer modified a RR (assuming that if they do this is because they want to use it to generate future replenishment) as Odoobot to make sure that only unmodified rules are deleted.